+++ ttl/cache/BUILD.bazel
@@ -54,18 +54,15 @@ go_test(
         "//parser/model",
         "//server",
         "//session",
-        "//store/helper",
+        "//store/mockstore",
         "//tablecodec",
         "//testkit",
         "//testkit/testsetup",
         "//ttl/session",
         "//types",
         "//util/codec",
-        "@com_github_pingcap_kvproto//pkg/metapb",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
-        "@com_github_tikv_client_go_v2//tikv",
-        "@com_github_tikv_pd_client//:client",
         "@org_uber_go_goleak//:goleak",
     ],
 )
